correccion materia-profesor

+++ Aulas/src/main/java/com/TrabajoFinal/Aulas/service/AulaService.java
@@ -1,6 +1,8 @@
 package com.TrabajoFinal.Aulas.service;
 
 import com.TrabajoFinal.Aulas.Repository.AulaRepository;
+import com.TrabajoFinal.Aulas.Repository.AvisoRepository;
+import com.TrabajoFinal.Aulas.Repository.ReservaRepository;
 import com.TrabajoFinal.Aulas.exceptions.ResourceNotFoundException;
 import com.TrabajoFinal.Aulas.model.Aula;
 import lombok.RequiredArgsConstructor;
@@ -12,6 +14,8 @@ import java.util.List;
 @RequiredArgsConstructor
 public class AulaService {
     private final AulaRepository aulaRepository;
+    private final ReservaRepository reservaRepository;
+    private final AvisoRepository avisoRepository;
 
     public List<Aula> listarAulas(){
         return aulaRepository.findAll();
@@ -28,7 +32,13 @@ public class AulaService {
 
     public void borrarAula(Integer id)  {
         Aula borrada = aulaXid(id);
-         aulaRepository.delete(borrada);
+        if (reservaRepository.existsByAulaId(id)) {
+            throw new RuntimeException("No se puede eliminar el aula porque tiene reservas asignadas.");
+        }
+        if (avisoRepository.existsByAulaId(id)) {
+            throw new RuntimeException("No se puede eliminar el aula porque tiene avisos pendientes.");
+        }
+        aulaRepository.delete(borrada);
     }
 
     public Aula modificarAula(Integer id, Aula nueva) {
